Update script and files, skel ignored by Calculate Utilities when created new user

+++ virtual/adrien-cldg-config/files/cldg-adrien.sh
@@ -1,29 +1,35 @@
 #! /bin/bash
 
-#Set french keyboard if language FR
-L=$(echo $LANG | cut -d_ -f1)
-K=$(gsettings get org.gnome.desktop.input-sources sources | cut -d\' -f4)
-
-if [[ "$L" == "fr" ]] && [[ "$K" == "us" ]]
+# First lanch of session
+if [ ! -e ~/.config/cldg-adrien.conf ]
 then
-	gsettings set org.gnome.desktop.input-sources sources "[('xkb', 'fr+oss')]"
+	#Set french keyboard if language FR
+	L=$(echo $LANG | cut -d_ -f1)
+	K=$(gsettings get org.gnome.desktop.input-sources sources | cut -d\' -f4)
+
+	if [[ "$L" == "fr" ]] && [[ "$K" == "us" ]]
+	then
+		gsettings set org.gnome.desktop.input-sources sources "[('xkb', 'fr+oss')]"
+	fi
+
+	# Background
+	gsettings set org.gnome.desktop.background picture-uri "'file:///usr/share/themes/Calculate/wallpaper.jpg'"
+
+	#Theme
+	gsettings set org.gnome.desktop.interface gtk-theme "Calculate"
+
+	# Some things if in live CD
+	root=$(df / | tail -1 | awk '{ print $1}')
+	if [[ "$root" == "none" ]]
+	then
+		# When icons on desktop
+		#DESKTOP=$(grep XDG_DESKTOP ~/.config/user-dirs.dirs | cut -d\" -f2)
+		#cp /etc/calculate/calculate-install.desktop $DESKTOP/
+		cp /etc/calculate/calculate-install.desktop /home/$USER/.local/share/applications/
+	fi
+
+	# Config file creation
+	echo "Initial CLDG configuration OK. To force the session to be customized, delete this file. Settings will be applied the next time you connect." >> ~/.config/cldg-adrien.conf
+	echo "Configuration initiale CLDG OK. Pour forcer la session à la personnalisation initiale, supprimer ce fichier. Le réglage sera appliqué à la prochaine connexion." >> ~/.config/cldg-adrien.conf
+	echo "Начальная конфигурация CLDG OK. Чтобы принудительно персонифицировать сеанс, удалите этот файл. Настройка будет применена при следующем подключении." >> ~/.config/cldg-adrien.conf
 fi
-
-# Background
-gsettings set org.gnome.desktop.background picture-uri "'file:///usr/share/themes/Calculate/wallpaper.jpg'"
-
-#Theme
-gsettings set org.gnome.desktop.interface gtk-theme "Calculate"
-
-# Some things if in live CD
-root=$(df / | tail -1 | awk '{ print $1}')
-if [[ "$root" == "none" ]]
-then
-	# When icons on desktop
-	#DESKTOP=$(grep XDG_DESKTOP ~/.config/user-dirs.dirs | cut -d\" -f2)
-	#cp /etc/calculate/calculate-install.desktop $DESKTOP/
-	cp /etc/calculate/calculate-install.desktop /home/$USER/.local/share/applications/
-fi
-
-
-rm -f ~/.config/autostart/cldg-adrien.desktop
